- Shift serialization, deserialization, and datagram code to its own .cpp and .h file
