+++ CloudyGameThinClient/Deserializer.h
@@ -1,0 +1,6 @@
+
+int deserialize_int8(unsigned char *buffer);
+int deserialize_int16(unsigned char *buffer);
+int deserialize_int32(unsigned char *buffer);
+
+PacketKeyboard deserialize_PacketKeyboard(unsigned char *buffer);
